Update up.sql
I have no recollection of ever touching this, but if it seems that I touched the nvarchar for [DisplayName] and changed it to 50. I looked back at the main dev repo and it's at 25 so I am changing it back to 25. I honestly do not remember touching this at all but just to double check